some changes

+++ sql_generator.py
@@ -1,9 +1,10 @@
+from os.path import join
+from collections import OrderedDict
+
 import pymysql
-from os.path import join, exists
 
 
 class ConnectionString:
-
     # host = '194.225.227.161'
     # port = 3306
     # user = 'leila_oskouie'
@@ -28,6 +29,10 @@ class SqlForeignKeyStructure(object):
         self.reference_column = reference_column
 
 
+def create_order_structure(table_structure, key_order):
+    return OrderedDict(sorted(table_structure.items(), key=lambda i: key_order.index(i[0])))
+
+
 def db_connection():
     try:
         connection = pymysql.connect(host=ConnectionString.host,
@@ -39,21 +44,26 @@ def db_connection():
                                      charset=ConnectionString.charset,
                                      )
         return connection
-    except Exception as e:
+    except Exception:
         return None
 
 
-def sql_create_table_command_generator(table_name, columns, primary_key=None, foreign_key=None, unique_key=None):
+def sql_create_table_command_generator(table_name, columns, primary_key=None, foreign_key=None,
+                                       unique_key=None, drop_table=False):
     """
     :param table_name:
     :param columns:
     :param primary_key: list of primary keys [key1, key2, ...]
-    :param foreign_key: list of dict [{'column':'col_name','reference_table':'table_name', 'reference_column':'col_name'},{}]
+    :param foreign_key: list of dict [{'column':'col_name','reference_table':'table_name',
+                                       'reference_column':'col_name'},{}]
     :param unique_key: defaultdict(list) {unique_name1: [col1, col2], unique_name2:[col1, col4]}
+    :param drop_table: boolean, set True for drop table command
     :return:
     """
-
-    command = "DROP TABLE IF EXISTS `%s`;\n" % table_name
+    if drop_table:
+        command = "DROP TABLE IF EXISTS `%s`;\n" % table_name
+    else:
+        command = ''
     command += "CREATE TABLE `%s` (\n " % table_name
 
     for key, value in columns.items():
@@ -61,7 +71,7 @@ def sql_create_table_command_generator(table_name, columns, primary_key=None, fo
     if primary_key:
         command += "PRIMARY KEY("
         for p_key in primary_key:
-            command += "`%s`," %p_key
+            command += "`%s`," % p_key
         command = command[:-1] + "),\n"
 
     if foreign_key:
@@ -79,32 +89,34 @@ def sql_create_table_command_generator(table_name, columns, primary_key=None, fo
     return command
 
 
-def insert_command(table_structure, table_name, insert_dictionary_columns, rows):
-
-    columns = ""
-    for name in insert_dictionary_columns:
-        columns += name+','
-    columns = columns[:-1]
-
-    command = "INSERT INTO `%s` (%s) VALUES " % (table_name, columns)
-
+def create_values_insert_command(rows, insert_columns, table_structure):
+    command = ''
     for records in rows:
         command += "( "
-
-        for column_name in insert_dictionary_columns:
+        for column_name in insert_columns:
             if 'varchar' in table_structure[column_name]:
                 command += " '%s', " % records[column_name].replace("'", "''").replace("\\", "\\\\")
             elif 'int' in table_structure[column_name]:
                 command += " %s, " % records[column_name]
-
         command = command[:-2] + '),'
-
     command = command[:-1] + ";"
+    return command
+
+
+def create_header_insert_command(insert_columns, table_name):
+    columns_name = '(`' + '`,`'.join(insert_columns) + '`)'
+    command = "INSERT INTO `%s`%s VALUES " % (table_name, columns_name)
+    return command
+
+
+def insert_command(table_structure, table_name, insert_columns, rows):
+    command = create_header_insert_command(insert_columns, table_name)
+    command += create_values_insert_command(rows, insert_columns, table_structure)
 
     return command
 
 
-def execute_command_mysql(command, message):
+def execute_command_mysql(command, message=None):
     conn = db_connection()
     if conn:
         try:
@@ -113,7 +125,7 @@ def execute_command_mysql(command, message):
             conn.commit()
 
         except Exception as e:
-            print('Some Exception Occur ' + str(e))
+            print('Some Exception Occurred ' + str(e))
             if message:
                 print('\n ' + message)
         finally:
@@ -127,57 +139,29 @@ def create_sql_dump(command, dir_path, file_name):
 
 
 def select_mysql(command):
-    try:
+    conn = db_connection()
+    if conn:
+        try:
+            cur = conn.cursor(pymysql.cursors.DictCursor)
+            result = cur.execute(command)
+            return result
 
-        conn = pymysql.connect(host=ConnectionString.host, port=ConnectionString.port, user=ConnectionString.user,
-                               passwd=ConnectionString.passwd,
-                               db=ConnectionString.db, use_unicode=ConnectionString.use_unicode,
-                               charset=ConnectionString.charset)
+        except Exception as e:
+            print('Some Exception Occurred' + str(e))
 
-        cur = conn.cursor(pymysql.cursors.DictCursor)
-        result = cur.execute(command)
-
-        return result
-
-    except Exception as e:
-        print('some exception occur ' + str(e))
-
-    finally:
-        conn.close()
+        finally:
+            conn.close()
 
 
 def select_result_mysql(command):
-    try:
-        conn = pymysql.connect(host=ConnectionString.host, port=ConnectionString.port, user=ConnectionString.user, passwd=ConnectionString.passwd,
-                               db=ConnectionString.db, use_unicode=ConnectionString.use_unicode, charset=ConnectionString.charset)
+    conn = db_connection()
+    if conn:
+        try:
+            cur = conn.cursor(pymysql.cursors.DictCursor)
+            cur.execute(command)
 
-        cur = conn.cursor(pymysql.cursors.DictCursor)
-        result = cur.execute(command)
-
-        return cur
-    except Exception as e:
-        print('some exception occur ' + str(e))
-    finally:
-        conn.close()
-
-
-def test():
-    key_order = ['id', 'template_name_fa', 'template_name_en', 'approved', 'col1', 'col2', 'col3']
-    table_structure = {
-        'id': 'int NOT NULL AUTO_INCREMENT',
-        'template_name_fa': 'varchar(500)',
-        'template_name_en': 'varchar(500)',
-        'approved': 'tinyint default NULL',
-        'col1': 'int',
-        'col2': 'int',
-        'col3': 'int'
-    }
-    table_name = 'test'
-    primary_keys = ['id', 'col1']
-    foreign_keys = [SqlForeignKeyStructure('col2', 'test2', 'id2'), SqlForeignKeyStructure('col3', 'test3', 'id3')]
-    unique_keys = {'col1_uniq': ['col1', 'col2'], 'col3_uniq': ['col2', 'col3']}
-    print(sql_create_table_command_generator(table_name, table_structure, primary_keys, foreign_keys, unique_keys))
-
-
-if __name__ == '__main__':
-    test()
+            return cur
+        except Exception as e:
+            print('Some Exception Occurred' + str(e))
+        finally:
+            conn.close()
